+++ test/libsolidity/semanticTests/array/copying/array_nested_memory_to_storage.sol
@@ -40,12 +40,12 @@ contract Test {
 // compileViaYul: also
 // ----
 // test() -> 24
-// gas irOptimized: 227167
+// gas irOptimized: 227133
 // gas legacy: 227133
 // gas legacyOptimized: 226547
 // test1() -> 3
 // test2() -> 6
 // test3() -> 24
-// gas irOptimized: 133621
+// gas irOptimized: 133597
 // gas legacy: 134295
 // gas legacyOptimized: 133383