+++ test/libsolidity/semanticTests/array/push/push_no_args_bytes.sol
@@ -23,7 +23,7 @@ contract C {
 // ----
 // l() -> 0
 // g(uint256): 70 ->
-// gas irOptimized: 184507
+// gas irOptimized: 185936
 // gas legacy: 184991
 // gas legacyOptimized: 180608
 // l() -> 70