+++ test/libsolidity/semanticTests/viaYul/copy_struct_invalid_ir_bug.sol
@@ -23,6 +23,6 @@ contract C {
 // compileViaYul: also
 // ----
 // f() ->
-// gas irOptimized: 113142
+// gas irOptimized: 112998
 // gas legacy: 112937
 // gas legacyOptimized: 112608