+++ test/libsolidity/syntaxTests/inlineAssembly/assembly_dialect_leading_space.sol
@@ -1,0 +1,5 @@
+function f() pure {
+    assembly " evmasm" {}
+}
+// ----
+// ParserError 4531: (33-42): Only "evmasm" supported.