+++ test/libsolidity/syntaxTests/viewPureChecker/access_to_base_members.sol
@@ -12,4 +12,4 @@ contract B is A {
 }
 // ----
 // TypeError 2527: (107-110): Function declared as pure, but this expression (potentially) reads from the environment or state and thus requires "view".
-// TypeError 8961: (157-160): Function declared as view, but this expression (potentially) modifies the state and thus requires non-payable (the default) or payable.
+// TypeError 8961: (157-160): Function cannot be declared as view because this expression (potentially) modifies the state.