+++ test/libsolidity/syntaxTests/viewPureChecker/creation_view_fail.sol
@@ -3,4 +3,4 @@ contract C {
     function f() public view { new D(); }
 }
 // ----
-// TypeError 8961: (58-65): Function declared as view, but this expression (potentially) modifies the state and thus requires non-payable (the default) or payable.
+// TypeError 8961: (58-65): Function cannot be declared as view because this expression (potentially) modifies the state.