+++ Changelog.md
@@ -13,13 +13,13 @@ Breaking changes:
 ### 0.8.13 (unreleased)
 
 Language Features:
+ * General: Allow annotating inline assembly as memory-safe to allow optimizations and stack limit evasion that rely on respecting Solidity's memory model.
 
 
 Compiler Features:
-
+* JSON-AST: Added selector field for errors and events.
 
 Bugfixes:
-
 
 
 ### 0.8.12 (2022-02-16)